multimodal-interview-agent/requirements.txt
# AI Interview Simulation Agent - Dependencies
# Python 3.8+ required

# Web Framework & API
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI & Machine Learning
openai>=1.3.0
openai-whisper>=20240930
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0

# Computer Vision & Proctoring
opencv-python>=4.8.0
mediapipe>=0.10.0
Pillow>=10.0.0

# Audio Processing & Speech
pyaudio>=0.2.11
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0
pyttsx3>=2.90

# Database & Storage
pymongo>=4.5.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Report Generation
reportlab>=4.0.0
matplotlib>=3.7.0

# Vector Search & RAG
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Core Utilities
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.1.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Development & Testing (Optional)
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0